+++ tests/xla_bridge_test.py
@@ -22,7 +22,7 @@ from absl.testing import absltest
 from jax._src import test_util as jtu
 from jax._src import xla_bridge as xb
 from jax._src.lib import xla_client as xc
-from jax.interpreters import xla
+from jax._src.interpreters import xla
 
 from jax._src.config import config
 config.parse_flags_with_absl()
